neue Texturen, Sprite angepasst
neue Textur für Explosion
neue Textur für Monstertod
Sprite leeren Konstruktor entfernt + float aus Render() durch int
ersetzt
Anpassungen überall, wo Sprite.Render(float ...) aufgerufen wurde

+++ Remarius/source/Explosion.cpp
@@ -30,7 +30,7 @@ void CExplosion::Render (float CameraX, float CameraY)														// Explosion
 	if (m_StateExplosion < 10)
 	{
 		m_pSpriteExplosion->SetPos (m_XPos - CameraX, m_YPos - CameraY);
-		m_pSpriteExplosion->Render (static_cast<float>(m_StateExplosion));
+		m_pSpriteExplosion->Render (m_StateExplosion);
 		if (m_StateExplosion > 3) m_bActive = false;
 	} else m_bIsAlive = false;
 }
